Add custom alert and confirm

+++ src/Species/wwwroot/js/Observation/Index/observation-info.js
@@ -55,9 +55,11 @@
         };
 
         vm.delete = () => {
-            vm.deleteObservation({ observation: vm.observation });
-            vm.$dismiss({
-                reason: 'delete',
+            $dialog.confirm('Вы действительно хотите удалить данное наблюдение?').then(() => {
+                vm.deleteObservation({ observation: vm.observation });
+                vm.$dismiss({
+                    reason: 'delete',
+                });
             });
         };
